move trpc setting from constants since it is not a constant
