+++ packages/frontend/src/hooks/useDbValues.ts
@@ -2,7 +2,7 @@ import { BulkKey, BulkKeyMap } from "@backend/utils/const";
 import { useQuery } from "@tanstack/react-query";
 import { createTRPCProxyClient } from "@trpc/client";
 import { AppRouter } from "@backend/index";
-import { trpcClientOptions } from "@/constants";
+import { trpcClientOptions } from "@/trpc";
 import { useAuth } from "./useAuth";
 
 const WORKER_RETRIEVAL_CHUNK_SIZE = 100;
